Move progress button
